requirements.txt
opencv-python>=4.10.0
pillow>=10.4.0
numpy>=1.26.4
pandas>=2.2.2
odfpy>=1.4.1
pytesseract>=0.3.13
pdf2image>=1.17.0
PyYAML>=6.0.1
ruamel.yaml>=0.17.0
PySide6>=6.6.0
pymupdf>=1.26.5

# dev only
pytest>=8.3.2
